add chat WebSocket
