+++ src/main/webapp/views/user/js/animations.js
@@ -1,7 +1,7 @@
 /* ========== Navbar show ========== */
 const navToggle = document.querySelector("#nav__toggle"),
-  navClose = document.querySelector("#nav__close"),
-  navMenu = document.querySelector("#nav__menu");
+    navClose = document.querySelector("#nav__close"),
+    navMenu = document.querySelector("#nav__menu");
 
 if (navToggle && navClose && navMenu) {
   navToggle.addEventListener("click", () => {
@@ -13,7 +13,7 @@ if (navToggle && navClose && navMenu) {
 }
 /* ========== carousel products ========== */
 const swiperContainer = document.querySelector(".swiper-container");
-if (swiperContainer) {  
+if (swiperContainer) {
   let carousel = new Swiper(".carousel", {
     loop: true,
     spaceBetween: 28,
@@ -36,16 +36,16 @@ if (swiperContainer) {
 }
 /* ========== Chat show ========== */
 const chatShow = document.querySelector("#chat__show"),
-  chatClose = document.querySelector("#chat__close"),
-  chat = document.querySelector("#chat"),
-  chatForm = document.querySelector("#chat__form");
+    chatClose = document.querySelector("#chat__close"),
+    chat = document.querySelector("#chat"),
+    chatForm = document.querySelector("#chat__form");
 
 
 if (chatShow && chatClose) {
   chatForm.addEventListener("submit", (e) => {
     e.preventDefault();
   });
-  
+
   chatShow.addEventListener("click", () => {
     chat.classList.toggle("chat__visible");
     document.body.classList.toggle("no-scroll");
@@ -53,36 +53,5 @@ if (chatShow && chatClose) {
   chatClose.addEventListener("click", () => {
     chat.classList.remove("chat__visible");
     document.body.classList.remove("no-scroll");
-  });
-
-  /* ========== writer in chat ========== */
-  const chatWriter = document.querySelector("#chat-writer"),
-    sendBtn = document.querySelector('#send-message'),
-    chatBox = document.querySelector('#chat-box');
-  
-  //create elements view message
-  const drawChatMessage = (message, type = 'request') => {
-    const boxMessage = document.createElement("LI"),
-    messageContent = document.createElement("SPAN");
-    
-    boxMessage.setAttribute("class", "chat__box");
-    messageContent.setAttribute("class", "chat__message");
-    boxMessage.appendChild(messageContent);
-
-    if (type !== 'request') type = 'response';
-    let typeMessageClass = 'chat__message--' + type;
-
-    boxMessage.classList.add(typeMessageClass);
-    messageContent.textContent = message;
-     
-    return boxMessage;
-  };
-
-  sendBtn.addEventListener("click", () => {
-    let message = chatWriter.value;
-  
-    chatBox.appendChild(drawChatMessage(message));
-    chatWriter.value = '';
-    chatWriter.focus();
   });
 }
